Move certain assets to dist/ root on build.
Certain asset files should exist at the root of a project (humans.txt, robots,txt, manifest.json, etc). This change will make webpack copy every file except those that shouldn't be copied to the asset folder, then it will copy only those files to the root folder of the build.

Perhaps these files could be refactored out into a more configurable area, such as the constants.js file, in case there are other needs for this, but these three are the primary offenders for this (in addition, sometimes, to the favicon).